requirements.txt
streamlit
transformers
torch
openai-whisper
pydub
